[Android] ✨Feat: 음식 사진 및 갤러리 구현  (S12P31S203-131)

See merge request s12-final/S12P31S203!69

+++ android-app/app/src/main/java/com/swallaby/foodon/data/main/remote/dto/MealRecordResponse.kt
@@ -10,9 +10,9 @@ data class MealRecordResponse(
     val imageUrl: String = "",
     val mealItems: List<String> = emptyList(),
     val totalKcal: Int = 0,
-    val totalCarbs: Int = 0,
-    val totalProtein: Int = 0,
-    val totalFat: Int = 0
+    val totalCarbs: Double = 0.0,
+    val totalProtein: Double = 0.0,
+    val totalFat: Double = 0.0,
 )
 
 fun MealRecordResponse.toDomain(): MealRecord {
